# ws/docs/openapi/provisioning.openapi.yaml
openapi: 3.1.0
info:
  title: Sukko Provisioning API
  version: 1.0.0
  description: |
    REST API for managing tenants, JWT signing keys, API keys, routing rules,
    channel rules, and quotas in the Sukko multi-tenant WebSocket platform.

    ## Authentication

    All `/api/v1/` endpoints require authentication via one of:
    - **Admin JWT**: `Authorization: Bearer {admin_jwt}` — Ed25519 keypair auth with `iss: "sukko-admin"`. Checked first; if present and valid, bypasses tenant JWT.
    - **Tenant JWT**: `Authorization: Bearer {jwt}` — validated against per-tenant asymmetric keys (ES256, RS256, EdDSA).

    Health, readiness, version, edition, config, and metrics endpoints (at root level, not under `/api/v1/`) require no authentication.

    ## Authorization

    Role-based access control with tenant isolation:
    - **admin/system**: Full access to all tenants and operations
    - **tenant owner**: Access to own tenant's resources only (JWT `tenant_id` must match)

    Operations like creating tenants, suspending, and modifying routing rules require `admin` or `system` role.

    ## Pagination

    All list endpoints support `?limit=N&offset=M` with defaults `limit=50, offset=0` (max limit=100).
    Responses include `{ items, total, limit, offset }`.

    ## Error Format

    All errors return: `{"code": "UPPER_SNAKE_CASE", "message": "human-readable"}`.
  contact:
    name: Sukko Platform Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/tenants/{tenantSlug}/webhooks:
    post:
      summary: Register a webhook endpoint
      operationId: createWebhook
      tags: [Webhooks]
      description: Creates a new webhook registration. Requires Pro edition. URL must use https (http allowed for local dev only).
      security:
        - TenantJWT: []
        - APIKey: []
      parameters:
        - $ref: "#/components/parameters/tenantSlug"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWebhookRequest"
      responses:
        "201":
          description: Webhook created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
        "400":
          description: Validation error (invalid URL, missing channel_pattern, max_retries out of range)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: WEBHOOK_INVALID_INPUT
                message: "webhook url must use https"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Feature not available in current edition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: EDITION_LIMIT
                message: "webhooks require Pro edition or higher"
        "422":
          description: Webhook quota exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: WEBHOOK_QUOTA_EXCEEDED
                message: "webhook quota exceeded"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      summary: List webhook registrations
      operationId: listWebhooks
      tags: [Webhooks]
      description: Returns paginated webhooks for the authenticated tenant.
      security:
        - TenantJWT: []
        - APIKey: []
      parameters:
        - $ref: "#/components/parameters/tenantSlug"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: Paginated webhook list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Feature not available in current edition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: EDITION_LIMIT
                message: "webhooks require Pro edition or higher"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/tenants/{tenantSlug}/webhooks/{webhookID}:
    get:
      summary: Get a webhook by ID
      operationId: getWebhook
      tags: [Webhooks]
      security:
        - TenantJWT: []
        - APIKey: []
      parameters:
        - $ref: "#/components/parameters/tenantSlug"
        - name: webhookID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Webhook found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Feature not available in current edition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: EDITION_LIMIT
                message: "webhooks require Pro edition or higher"
        "404":
          description: Webhook not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: WEBHOOK_NOT_FOUND
                message: "webhook not found"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      summary: Partially update a webhook
      operationId: updateWebhook
      tags: [Webhooks]
      description: All fields are optional. Omitted fields are unchanged. Status transitions — enabled ↔ degraded ↔ suspended.
      security:
        - TenantJWT: []
        - APIKey: []
      parameters:
        - $ref: "#/components/parameters/tenantSlug"
        - name: webhookID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWebhookRequest"
      responses:
        "200":
          description: Updated webhook
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Feature not available in current edition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: EDITION_LIMIT
                message: "webhooks require Pro edition or higher"
        "404":
          description: Webhook not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete a webhook
      operationId: deleteWebhook
      tags: [Webhooks]
      security:
        - TenantJWT: []
        - APIKey: []
      parameters:
        - $ref: "#/components/parameters/tenantSlug"
        - name: webhookID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Webhook deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Feature not available in current edition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: EDITION_LIMIT
                message: "webhooks require Pro edition or higher"
        "404":
          description: Webhook not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Admin JWT (Ed25519 keypair, iss:"sukko-admin") or tenant JWT (per-tenant asymmetric keys).
        Admin JWT is checked first; if the issuer is not "sukko-admin", tenant JWT validation is used.
  parameters:
    tenantSlug:
      name: tenantSlug
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z][a-z0-9-]{2,62}$"
      description: Tenant slug — the human-readable Kafka namespace and URL identifier (3–63 chars, lowercase alphanumeric + hyphens, must start with a letter)
...
    limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 100
      description: Number of items to return
    offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
      description: Number of items to skip
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code (UPPER_SNAKE_CASE)
        message:
          type: string
          description: Human-readable error description
...
    Webhook:
      type: object
      required: [id, tenant_id, url, channel_pattern, status, max_retries, retry_count, created_at, updated_at]
      properties:
        id:
          type: string
          description: "Webhook ID (wh_ prefix + base32)"
          example: wh_abc1def2ghi3
        tenant_id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
          description: HTTPS endpoint URL
        channel_pattern:
          type: string
          description: Channel glob pattern (e.g. trades.*)
        status:
          type: string
          enum: [enabled, degraded, suspended]
        max_retries:
          type: integer
          minimum: 1
          maximum: 10
          description: Maximum delivery attempts before entering degraded status
        retry_count:
          type: integer
          description: Number of consecutive delivery failures since last success
        last_status:
          type: ["string", "null"]
          description: HTTP status code string of last delivery attempt (null if never delivered)
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateWebhookRequest:
      type: object
      required: [url, channel_pattern, secret]
      properties:
        url:
          type: string
          format: uri
          description: Delivery endpoint. Must be https (http allowed only when WEBHOOK_ALLOW_HTTP=true in local environments).
        channel_pattern:
          type: string
          description: Channel glob pattern that triggers this webhook (e.g. trades.*)
        secret:
          type: string
          description: Plaintext HMAC secret. Encrypted at rest; never returned in API responses.
        max_retries:
          type: integer
          minimum: 1
          maximum: 10
          default: 5
          description: Maximum delivery attempts before entering degraded status (0 = use default of 5)
    UpdateWebhookRequest:
      type: object
      description: All fields are optional. Omitted fields are unchanged.
      properties:
        url:
          type: string
          format: uri
        channel_pattern:
          type: string
        max_retries:
          type: integer
          minimum: 1
          maximum: 10
        status:
          type: string
          enum: [enabled, degraded, suspended]
          description: Explicit status transition. Transitioning to 'enabled' resets retry_count to 0.
    WebhookListResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Webhook"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
...